LPD-62043 Refactor on Trial form, and autou pdate table
