+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/components/ListView/ListView.tsx
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {ClayPaginationBarWithBasicItems} from '@clayui/pagination-bar';
+import { ClayPaginationBarWithBasicItems } from '@clayui/pagination-bar';
 import React, {
 	ComponentProps,
 	ReactNode,
@@ -12,11 +12,11 @@ import React, {
 	useEffect,
 	useMemo,
 } from 'react';
-import {useSearchParams} from 'react-router-dom';
-import {KeyedMutator} from 'swr';
+import { useSearchParams } from 'react-router-dom';
+import { KeyedMutator } from 'swr';
 
 import CreateFilters from '../../core/CreateFilters';
-import {useFetch} from '../../hooks/useFetch';
+import { useFetch } from '../../hooks/useFetch';
 import i18n from '../../i18n';
 import {
 	FilterSchema as FilterSchemaType,
@@ -30,7 +30,7 @@ import Loading from '../Loading';
 import ManagementToolbar, {
 	ManagementToolbarProps,
 } from './components/ManagementToolbar';
-import Table, {TableProps} from './components/Table';
+import Table, { TableProps } from './components/Table';
 import ListViewContextProvider, {
 	AppActions,
 	InitialState as ListViewContextState,
@@ -53,7 +53,7 @@ export type ListViewProps<T extends Record<string, any>> = {
 		options: ChildrenOptions
 	) => ReactNode;
 
-	defaultFilters?: {filter: string};
+	defaultFilters?: { filter: string };
 
 	emptyStateProps?: ComponentProps<typeof EmptyState>;
 
@@ -80,12 +80,14 @@ export type ListViewProps<T extends Record<string, any>> = {
 
 	/**
 	 * The options for the pagination.
-	 *
-	 * @default {displayType: true}
-	 */
+	*
+	* @default {displayType: true}
+	*/
 	paginationOptions?: {
 		displayType: boolean;
 	};
+
+	refreshInterval?: number;
 
 	resource: string;
 
@@ -123,17 +125,19 @@ const ListView = <T extends Record<string, any>>({
 		visible: managementToolbarVisible = false,
 		...managementToolbarProps
 	} = {},
-	paginationOptions = {displayType: true},
+	paginationOptions = { displayType: true },
 	resource,
 	tableProps,
 	transformData = (item) => item,
+	refreshInterval
 }: ListViewProps<T>) => {
 	const [listViewContext, dispatch] = useContext(ListViewContext);
 
 	const updateUrlParams = useUpdateUrlParams();
 	const [searchParams] = useSearchParams();
 
-	const {filters, keywords, sort} = listViewContext;
+	const { filters, keywords, sort } = listViewContext;
+	const filterSchemaName = managementToolbarProps?.filterSchema ?? '';
 
 	const filterSchema = (filterSchemas as any)[
 		managementToolbarProps?.filterSchema ?? ''
@@ -204,7 +208,7 @@ const ListView = <T extends Record<string, any>>({
 	const filter = useMemo(() => {
 		const baseFilter = CreateFilters.createFilter(filterVariables) || '';
 
-		return {filter: baseFilter};
+		return { filter: baseFilter };
 	}, [filterVariables]);
 
 	const buildSort = (sort: Sort) =>
@@ -213,7 +217,7 @@ const ListView = <T extends Record<string, any>>({
 	const onSort = useCallback(
 		(key: string, direction: SortDirection) => {
 			dispatch({
-				payload: {direction, key},
+				payload: { direction, key },
 				type: ListViewTypes.SET_SORT,
 			});
 		},
@@ -249,7 +253,7 @@ const ListView = <T extends Record<string, any>>({
 		mutate,
 	} = useFetch(resource, {
 		params: getURLSearchParams(),
-	});
+	}, refreshInterval);
 
 	const {
 		actions = {},
@@ -257,7 +261,7 @@ const ListView = <T extends Record<string, any>>({
 		page = 1,
 		pageSize,
 		totalCount = 0,
-	} = transformData(response || {items: []});
+	} = transformData(response || { items: [] });
 
 	if (loading || (isValidating && searchParams.get('filter'))) {
 		return <Loading />;
@@ -277,14 +281,14 @@ const ListView = <T extends Record<string, any>>({
 				selectPerPageItems: i18n.translate('x-items'),
 			}}
 			onDeltaChange={(delta) => {
-				updateUrlParams({pageSize: delta});
+				updateUrlParams({ pageSize: delta });
 
-				dispatch({payload: delta, type: ListViewTypes.SET_PAGE_SIZE});
+				dispatch({ payload: delta, type: ListViewTypes.SET_PAGE_SIZE });
 			}}
 			onPageChange={(page) => {
-				updateUrlParams({page});
+				updateUrlParams({ page });
 
-				dispatch({payload: page, type: ListViewTypes.SET_PAGE});
+				dispatch({ payload: page, type: ListViewTypes.SET_PAGE });
 			}}
 			totalItems={totalCount}
 		/>
